Mejoras y correcciones a la pagina

+++ SistemaEmisiones/codes/js/session.js
@@ -1,22 +1,19 @@
 const form = document.getElementById('form');
 const alertMessage = document.getElementById('alert');
 
-//
-function getCookie(cookie){
-
-}
 
 document.addEventListener("DOMContentLoaded", () => {
-  alertMessage.style.display ="none";
+  if(alertMessage != null){
+    alertMessage.style.display ="none";
   
-  if(sessionStorage.getItem('error') == 'error'){
-    sessionStorage.clear();
-    alertMessage.style.display =  "block";
-  }
+    if(sessionStorage.getItem('error') == 'error'){
+      sessionStorage.clear();
+      alertMessage.style.display =  "block";
+    }
+  }  
 });
 
 form.addEventListener('submit', (e)=>{
   sessionStorage.setItem('error', 'error');
 })
-
